Add IGenericRepository interface for entity management
Introduces the `IGenericRepository<TEntity>` interface in the `AuthServer.Core.Repositories` namespace. This interface implements a generic repository pattern, providing methods for asynchronous retrieval, querying, adding, updating, and removing entities of type `TEntity`. The `System.Linq.Expressions` namespace is imported to facilitate the use of expression trees in the `Where` method.